requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Redis
redis==5.0.1
hiredis==2.2.3

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Audio processing
librosa==0.10.1
soundfile==0.12.1
numpy>=1.26.0  # Python 3.12 compatible

# ML/AI
torch>=2.1.0
torchaudio>=2.1.0
scikit-learn>=1.3.2
pandas>=2.1.3

# Text processing
nltk==3.8.1
transformers==4.35.2
datasets>=2.14.0  # For GoEmotions and other Hugging Face datasets

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
python-multipart==0.0.6
pyyaml==6.0.1

